requirements.txt
Flask==1.1.4
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
pyjwt==1.7.1
Flask-Script==2.0.6
Flask-Migrate==2.5.3
markupsafe==2.0.1
six==1.16.0
gunicorn
flask_restplus
werkzeug==0.16.1
requests==2.28.1


